Rename built_in_sensor_types to built_in_content
